Test Refactor - DefaultTestContext
- Bring helper classes into same file
- Reinforces that @DefaultTestContext is what to use

[nostory]
