+++ uaa/src/test/java/org/cloudfoundry/identity/uaa/DefaultTestContext.java
@@ -1,11 +1,20 @@
 package org.cloudfoundry.identity.uaa;
 
 import org.cloudfoundry.identity.uaa.extensions.PollutionPreventionExtension;
+import org.cloudfoundry.identity.uaa.test.TestClient;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ImportResource;
+import org.springframework.context.annotation.PropertySource;
+import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
+import org.springframework.security.web.FilterChainProxy;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.context.web.WebAppConfiguration;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import org.springframework.web.context.WebApplicationContext;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -23,4 +32,32 @@ import java.lang.annotation.Target;
         TestClientAndMockMvcTestConfig.class,
 })
 public @interface DefaultTestContext {
+}
+
+@ImportResource(locations = {"file:./src/main/webapp/WEB-INF/spring-servlet.xml"})
+@PropertySource(value = "classpath:integration_test_properties.yml", factory = NestedMapPropertySourceFactory.class)
+class SpringServletTestConfig {
+    @Bean
+    public static PropertySourcesPlaceholderConfigurer properties() {
+        return new PropertySourcesPlaceholderConfigurer();
+    }
+}
+
+class TestClientAndMockMvcTestConfig {
+    @Bean
+    public MockMvc mockMvc(
+            WebApplicationContext webApplicationContext,
+            @SuppressWarnings("SpringJavaInjectionPointsAutowiringInspection") FilterChainProxy springSecurityFilterChain
+    ) {
+        return MockMvcBuilders.webAppContextSetup(webApplicationContext)
+                .addFilter(springSecurityFilterChain)
+                .build();
+    }
+
+    @Bean
+    public TestClient testClient(
+            MockMvc mockMvc
+    ) {
+        return new TestClient(mockMvc);
+    }
 }
