--- uaa/src/test/java/org/cloudfoundry/identity/uaa/TestClientAndMockMvcTestConfig.java
@@ -1,27 +1,0 @@
-package org.cloudfoundry.identity.uaa;
-
-import org.cloudfoundry.identity.uaa.test.TestClient;
-import org.springframework.context.annotation.Bean;
-import org.springframework.security.web.FilterChainProxy;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.web.context.WebApplicationContext;
-
-public class TestClientAndMockMvcTestConfig {
-    @Bean
-    public MockMvc mockMvc(
-            WebApplicationContext webApplicationContext,
-            @SuppressWarnings("SpringJavaInjectionPointsAutowiringInspection") FilterChainProxy springSecurityFilterChain
-    ) {
-        return MockMvcBuilders.webAppContextSetup(webApplicationContext)
-                .addFilter(springSecurityFilterChain)
-                .build();
-    }
-
-    @Bean
-    public TestClient testClient(
-            MockMvc mockMvc
-    ) {
-        return new TestClient(mockMvc);
-    }
-}
